SPARKC-481: Refactor Custom Scan Method
The original implmentation of the CustomScanMethod made it difficult to
extend the functionality inside of CassandraTableScan RDD since it
provided no ability to perform a "close" action on termination. This
meant external users were unable to properly cleanup their
implementations.

To fix this we change getScanMethod to getScanner which returns a
Scanner object which provides two methods, Scan and Close. One Scanner
is built per Spark Partition so the the initializer will only be called
once per Spark Partition and the close method only once at task
Completion.